+++ src/assets/css/sidebar.scss
@@ -50,78 +50,6 @@
           align-items: center;
         }
       }
+      
     }
 }
-// nav{
-//   width: 100%;
-//   max-width: 400px;
-//   position: relative;
-//   border-right: 1px solid #9296A6;
-//   background: #ffffff;
-//   transition: 0.3s;
-//   min-height: 100vh;
-
-//   &.active {
-//     width: 80px;
-//     .logo {
-//       width: 0px;
-//       overflow: hidden;
-//       transition: 0.3s;
-//     }
-//     .wrapper{
-//       .item {
-//         margin: 0px 0px;
-
-//         span {
-//           width: 0px;
-//           overflow: hidden;
-//           display: block;
-//           height: 20px;
-//           transition: 0.3s;
-//         }
-//       }
-//     }
-//     .logo-wrapper {
-//       padding: 0px;
-//       i {
-//         width: 100%;
-//       }
-//     }
-//   }
-//   .nav-main-block{
-//     height: calc(100% - 80px);
-//   }
-//   .logo {
-//     width: 80%;
-//     overflow: hidden;
-//     transition: 0.3s;
-//   }
-
-//   .logo-wrapper{
-//     display: flex;
-//     justify-content: center;
-//     align-items: center;
-//     height: 80px;
-//     padding-right: 10px;
-//     transition: 0.3s;
-
-//     i{
-//       width: 20%;
-//       font-size: 34px;
-//       font-weight: 100;
-//       color: #24af42;
-//       justify-content: center;
-//       align-items: center;
-//       display: flex;
-
-//     }
-//   }
-// }
-
-// @media screen and (max-width:980px) { /* Максимум ширина 980px */
-//   nav {
-//     left: -302px;
-//     position: absolute;
-//   }
-
-// }
