+++ src/assets/css/variables.scss
@@ -5,6 +5,7 @@ $greyOne: rgba(146, 150, 166, 1);
 $greyTwo: rgba(215, 215, 219, 1);
 $greyFour: rgba(244, 244, 246, 1);
 $white: rgba(255,255,255,1);
+$blackOne: rgba(0, 0, 0, 1);
 $blackTwo: rgba(15, 27, 56, 1);
 $activeColor: rgba(36,183,66,1);
 $todoColor: #002f6d;
