+++ src/components/app/FooterNavbar.vue
@@ -1,12 +1,10 @@
 <template>
   <div class="footer-nav">
     <div class="wrapper">
-      <div class="copyright" v-if="!actives">
+      <div class="copyright">
         <span>2022&#169; Transporter
           <br />
         Все права защищены</span>
-
-        <!-- <a href="">Противодействие мошейничеству корупции и хищениям</a> -->
       </div>
     </div>
   </div>
@@ -15,22 +13,6 @@
 <script>
 export default {
   name: "FooterNavbar",
-  props: {
-    actives : {
-      type: Boolean,
-    }
-  },
-  components: {
-  },
-  methods: {
-    setLocale(locale){
-      console.log(locale);
-      import(`@/langs/${locale}.json`).then((msg) => {
-        this.$i18n.setLocaleMessage(locale , msg)
-        this.$i18n.locale = locale
-      })
-    }
-  },
 }
 </script>
