+++ src/langs/en.json
@@ -16,5 +16,20 @@
   "settings" : "Settings",
   "language" : "language",
   "TheFolderIsEmpty" : "The folder is empty",
-  "Current" : "Current"
+  "Current" : "Current",
+  "Settings" : {
+    "Title" : "Settings",
+    "Project" : "Project Settings",
+    "Users" : "Profile Settings",
+    "TelegramAlert" : "Notification Settings",
+    "ForgeAPI" : "Forge Api",
+    "project" : {
+      "UpdateProjectList" : "Update Project List",
+      "ProjectTitle" : "Project Title",
+      "Synchronization" : "Synchronization",
+      "LastDetour" : "Last detour",
+      "Progress" : "Progress"
+    }
+    
+  }
 }
