+++ src/langs/ru.json
@@ -16,6 +16,21 @@
   "settings" : "Настройки",
   "language" : "Язык",
   "TheFolderIsEmpty" : "Папка пустая",
-  "Current" : "Текущая"
+  "Current" : "Текущая",
+  "Settings" : {
+    "Title" : "Настройки",
+    "Project" : "Настройки проектов",
+    "Users" : "Настройки профиля",
+    "TelegramAlert" : "Настройки уведомления",
+    "ForgeAPI" : "Forge Api",
+    "project" : {
+      "UpdateProjectList" : "Обновить список проектов",
+      "ProjectTitle" : "Название проекта",
+      "Synchronization" : "Синхронизация",
+      "LastDetour" : "Последний обход",
+      "Progress" : "Прогресс"
+
+    }
+  }
 
 }
