+++ src/router/index.js
@@ -17,6 +17,30 @@ const routes = [
     component: () => import('../views/Login.vue'),
   },
   {
+    path: '/forgeAPISetting',
+    name: 'forgeAPISetting',
+    meta: {layout: 'Setting'},
+    component: () => import('../views/settings/ForgeAPISetting.vue'),
+  },
+  {
+    path: '/telegramAlertSetting',
+    name: 'telegramAlertSetting',
+    meta: {layout: 'Setting'},
+    component: () => import('../views/settings/TelegramAlertSetting.vue'),
+  },
+  {
+    path: '/projectSetting',
+    name: 'projectSetting',
+    meta: {layout: 'Setting'},
+    component: () => import('../views/settings/ProjectSetting.vue'),
+  },
+  {
+    path: '/userSetting',
+    name: 'userSetting',
+    meta: {layout: 'Setting'},
+    component: () => import('../views/settings/UserSetting.vue'),
+  },
+  {
     path: '/list',
     name: 'list',
     meta: {layout: 'main'},
