+++ src/views/settings/UserSetting.vue
@@ -1,0 +1,18 @@
+<template>
+<div>dfdfd</div>
+</template>
+<script>
+// import UiButton from '@/components/ui/Button'
+// import UiInput from '@/components/ui/Input'
+// import btnSec from '@/components/ui/BtnSec'
+// import UiMark from '@/components/ui/Mark'
+// import UiTabs from '@/components/ui/Tabs'
+// import {HTTP} from "@/http-common";
+
+export default {
+}
+</script>
+<style lang="scss">
+
+@import 'src/assets/css/users-settings.scss';
+</style>
